requirements.txt
# Core dependencies
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# PDF processing
PyPDF2>=3.0.0
pdfplumber>=0.10.0

# AI/OpenAI
openai>=1.3.0

# Google Sheets
gspread>=5.12.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1

# Telegram
python-telegram-bot>=20.7

# Scheduling
schedule>=1.2.0

# Environment variables
python-dotenv>=1.0.0

# HTML parsing
html5lib>=1.1

# Rate limiting
ratelimit>=2.2.1

# Logging
colorlog>=6.8.0

# Date/Time utilities
python-dateutil>=2.8.2

# JSON handling
jsonschema>=4.20.0

# Retry logic
tenacity>=8.2.3
